+++ intasend/src/intasend/mod.rs
@@ -1,12 +1,13 @@
 #![allow(unused)]
 #![allow(unused_imports)]
 
-use anyhow::{Error as StdErr, Result};
-use reqwest::{Client, Error};
+use anyhow::Result;
+use reqwest::{Client, Error as ReqwestErr};
 use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use serde_json::Value as JSON;
-use std::fmt::Debug;
+use std::fmt::{self, Debug};
+use thiserror::Error as ThisErr;
 
 #[cfg(feature = "client")]
 #[cfg(feature = "server")]
@@ -16,18 +17,18 @@ pub(crate) mod payouts;
 pub(crate) mod refunds;
 pub(crate) mod wallets;
 
-use checkout::Checkout;
-use collection::Collection;
-use payouts::Payouts;
-use refunds::Refunds;
-use wallets::Wallets;
+use checkout::CheckoutsAPI;
+use collection::CollectionsAPI;
+use payouts::PayoutsAPI;
+use refunds::RefundsAPI;
+use wallets::WalletsAPI;
 
 /// **[IntaSend](https://intasend.com)** - The _Unoffical_ Rust Client SDK for the Intasend API Gateway.
 ///
 /// This library is a wrapper around the IntaSend Payment Gateway that supports a
 /// variety of payment methods e.g Visa, Mastercard, M-Pesa, and even Bitcoin.
 ///
-/// The library is fully async and it uses Reqwest library under the hood to make asynchronous calls to the REST API.  
+/// The library is fully async and it uses Reqwest library under the hood to make asynchronous calls to the REST API.
 ///
 /// To use the library you should acquire test or production API keys here: [Sandbox](https://sandbox.intasend.com) or [Production](https://payment.intasend.com)
 ///
@@ -76,8 +77,8 @@ impl Intasend {
     ///
     /// ```
     #[cfg(feature = "server")]
-    pub fn collection(&self) -> Collection {
-        Collection {
+    pub fn collection(&self) -> CollectionsAPI {
+        CollectionsAPI {
             intasend: self.clone(),
         }
     }
@@ -91,8 +92,8 @@ impl Intasend {
     ///
     /// ```
     #[cfg(feature = "client")]
-    pub fn checkout(&self) -> Checkout {
-        Checkout {
+    pub fn checkout(&self) -> CheckoutsAPI {
+        CheckoutsAPI {
             intasend: self.clone(),
         }
     }
@@ -106,8 +107,8 @@ impl Intasend {
     ///
     /// ```
     #[cfg(feature = "server")]
-    pub fn payouts(&self) -> Payouts {
-        Payouts {
+    pub fn payouts(&self) -> PayoutsAPI {
+        PayoutsAPI {
             intasend: self.clone(),
         }
     }
@@ -121,8 +122,8 @@ impl Intasend {
     ///
     /// ```
     #[cfg(feature = "server")]
-    pub fn refunds(&self) -> Refunds {
-        Refunds {
+    pub fn refunds(&self) -> RefundsAPI {
+        RefundsAPI {
             intasend: self.clone(),
         }
     }
@@ -136,8 +137,8 @@ impl Intasend {
     ///
     /// ```
     #[cfg(feature = "server")]
-    pub fn wallets(&self) -> Wallets {
-        Wallets {
+    pub fn wallets(&self) -> WalletsAPI {
+        WalletsAPI {
             intasend: self.clone(),
         }
     }
@@ -154,7 +155,7 @@ impl RequestClient for Intasend
         payload: Option<T>,
         service_path: &str,
         request_method: RequestMethods,
-    ) -> Result<U, StdErr> {
+    ) -> Result<U, IntasendClientError> {
         let client = Client::new();
 
         let base_url = if self.test_mode {
@@ -164,7 +165,7 @@ impl RequestClient for Intasend
         };
 
         match request_method {
-            RequestMethods::GET => {
+            RequestMethods::Get => {
                 let response = client
                     .get(&format!("{}{}", base_url, service_path))
                     .header("Content-Type", "application/json")
@@ -180,7 +181,7 @@ impl RequestClient for Intasend
 
                 Ok(json)
             }
-            RequestMethods::POST => {
+            RequestMethods::Post => {
                 let response = client
                     .post(&format!("{}{}", base_url, service_path))
                     .header("Content-Type", "application/json")
@@ -206,7 +207,7 @@ impl RequestClient for Intasend
         payload: Option<T>,
         service_path: &str,
         request_method: RequestMethods,
-    ) -> Result<U, StdErr> {
+    ) -> Result<U, IntasendClientError> {
         let client = Client::new();
 
         let base_url = if self.test_mode {
@@ -216,42 +217,64 @@ impl RequestClient for Intasend
         };
 
         match request_method {
-            RequestMethods::GET => {
+            RequestMethods::Get => {
                 let response = client
                     .get(&format!("{}{}", base_url, service_path))
                     .header("Content-Type", "application/json")
                     .header("Authorization", format!("Bearer {}", self.secret_key))
-                    .header("X-IntaSend-Public-API-Key", self.publishable_key.clone())
+                    // .header("X-IntaSend-Public-API-Key", self.publishable_key.clone())
                     .send()
-                    .await;
-                // println!("[#] API Response: {:#?}", response);
+                    .await?;
+                // .json()
+                // .await?;
+                // println!("[#] API Response: {:?}", response);
 
                 // let json: Map<String, Value> = serde_json::from_str(response)?;
-                let json = serde_json::from_value::<U>(response?.json().await?)
-                    .expect("[!] Error parsing json!");
+                // let json = serde_json::from_value::<U>(response?.json().await?)
+                //     .expect("[!] Error parsing json!");
 
-                Ok(json)
+                if response.status().is_success() {
+                    let parsed_response = response.json::<U>().await?;
+                    // println!("[#] API Response (OK): {:#?}", parsed_response);
+                    Ok(parsed_response)
+                } else {
+                    Err(IntasendClientError::UnexpectedResponseStatus(
+                        response.status(),
+                    ))
+                }
+
+                // Ok(response)
             }
-            RequestMethods::POST => {
-                let response: U = client
+            RequestMethods::Post => {
+                let response = client
                     .post(&format!("{}{}", base_url, service_path))
                     .header("Content-Type", "application/json")
                     .header("Authorization", format!("Bearer {}", self.secret_key))
                     // .header("X-IntaSend-Public-API-Key", self.publishable_key.clone())
                     .json(&payload)
                     .send()
-                    .await?
-                    .json()
                     .await?;
-                println!("[#] API Response: {:#?}", response);
+                    // .json()
+                    // .await?;
+                // println!("[#] API Response: {:#?}", response);
 
                 // let json: Map<String, Value> = serde_json::from_str(response)?;
                 // let json = serde_json::from_value::<U>(response?.json().await?)
                 //     .expect("[!] Error parsing json!");
                 // let json = serde_json::from_str::<U>(&response).expect("[!] Error parsing json!");
-
                 // Ok(json)
-                Ok(response)
+
+                if response.status().is_success() {
+                    let parsed_response: U = response.json().await?;
+                    // println!("[#] API Response (OK): {:#?}", parsed_response);
+                    Ok(parsed_response)
+                } else {
+                    Err(IntasendClientError::UnexpectedResponseStatus(
+                        response.status(),
+                    ))
+                }
+
+                // Ok(response)
             }
         }
     }
@@ -263,33 +286,52 @@ pub trait RequestClient {
         payload: Option<T>,
         service_path: &str,
         request_method: RequestMethods,
-    ) -> Result<U, StdErr>;
+    ) -> Result<U, IntasendClientError>;
     async fn send<T: Serialize, U: for<'de> Deserialize<'de> + Debug>(
         &self,
         payload: Option<T>,
         service_path: &str,
         request_method: RequestMethods,
-    ) -> Result<U, StdErr>;
+    ) -> Result<U, IntasendClientError>;
 }
 
-#[derive(Debug)]
-pub enum RequestClientError {
-    ReqwestError(reqwest::Error),
-    SerdeJsonError(serde_json::Error),
+// #[derive(Error, Debug)]
+// pub enum DataStoreError {
+//     #[error("data store disconnected")]
+//     Disconnect(#[from] io::Error),
+//     #[error("the data for key `{0}` is not available")]
+//     Redaction(String),
+//     #[error("invalid header (expected {expected:?}, found {found:?})")]
+//     InvalidHeader {
+//         expected: String,
+//         found: String,
+//     },
+//     #[error("unknown data store error")]
+//     Unknown,
+// }
+
+#[derive(ThisErr, Debug)]
+pub enum IntasendClientError {
+    #[error("Intasend client error")]
+    ReqwestError(#[from] ReqwestErr),
+    #[error("Intasend json error")]
+    SerdeJsonError(#[from] serde_json::Error),
+    #[error("Unexpected response status: {0}")]
+    UnexpectedResponseStatus(reqwest::StatusCode),
     // ... other error variants
 }
 
-impl From<reqwest::Error> for RequestClientError {
-    fn from(err: reqwest::Error) -> Self {
-        RequestClientError::ReqwestError(err)
-    }
-}
+// impl From<reqwest::Error> for IntasendClientError {
+//     fn from(err: reqwest::Error) -> Self {
+//         IntasendClientError::ReqwestError(err)
+//     }
+// }
 
-impl From<serde_json::Error> for RequestClientError {
-    fn from(err: serde_json::Error) -> Self {
-        RequestClientError::SerdeJsonError(err)
-    }
-}
+// impl From<serde_json::Error> for IntasendClientError {
+//     fn from(err: serde_json::Error) -> Self {
+//         IntasendClientError::SerdeJsonError(err)
+//     }
+// }
 
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub struct Invoice {
@@ -422,8 +464,8 @@ impl Provider {
 }
 
 pub enum RequestMethods {
-    GET,
-    POST,
+    Get,
+    Post,
 }
 
 // Define the trait for JSON conversion
@@ -457,7 +499,14 @@ impl Currency {
     }
 }
 
+impl fmt::Display for Currency {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
+}
+
 #[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum Tarrif {
     BusinessPays,
     CustomerPays,
@@ -481,6 +530,7 @@ impl Tarrif {
 
 /// Payout Provider Options supported by Intasend API Gateway
 #[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "SCREAMING_SNAKE_CASE")]
 pub enum PayoutProvider {
     MpesaB2c,
     MpesaB2b,