+++ intasend/src/lib.rs
@@ -3,16 +3,20 @@
 mod intasend;
 
 pub use self::intasend::checkout::{
-    Checkout, CheckoutDetailsRequest, CheckoutDetailsResponse, CheckoutMethod, CheckoutRequest,
-    CheckoutResponse,
+    CheckoutDetailsRequest, CheckoutDetailsResponse, CheckoutMethod, CheckoutRequest,
+    CheckoutResponse, CheckoutsAPI,
 };
 pub use self::intasend::collection::{
-    Collection, MpesaStkPushRequest, MpesaStkPushResponse, StkPushStatusRequest, StkPushStatusResponse,
+    CollectionsAPI, MpesaStkPushRequest, MpesaStkPushResponse, StkPushStatusRequest,
+    StkPushStatusResponse,
 };
-pub use self::intasend::payouts::{Payout, PayoutTransaction, PayoutRequest, Payouts};
-pub use self::intasend::refunds::{RefundRequest, Refunds, RefundResponse};
-pub use self::intasend::wallets::{FundCheckoutRequest, Wallet, Wallets};
-pub use self::intasend::{Currency, Intasend, Transaction, PayoutProvider};
+pub use self::intasend::payouts::{Payout, PayoutRequest, PayoutTransaction, PayoutsAPI};
+pub use self::intasend::refunds::{RefundRequest, RefundResponse, RefundsAPI};
+pub use self::intasend::wallets::{
+    FundCheckoutRequest, Wallet, WalletCreateDetails, WalletDetailsRequest, WalletListResponse,
+    WalletType, WalletsAPI,
+};
+pub use self::intasend::{Currency, Intasend, PayoutProvider, Transaction};
 
 #[cfg(test)]
 mod tests {